zoom works!!!!!

+++ fractol.h
@@ -6,7 +6,7 @@
 /*   By: zhabri <zhabri@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/02 08:44:47 by zhabri            #+#    #+#             */
-/*   Updated: 2022/11/07 19:39:36 by zhabri           ###   ########.fr       */
+/*   Updated: 2022/11/08 11:53:58 by zhabri           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,43 +29,43 @@ typedef struct s_pixel
 	int			x;
 	int			y;
 	int			color;
-}			t_pixel;
+}				t_pixel;
 
 typedef struct s_image
 {
-	void	*img;
-	char	*addr;
-	int		bpp;
-	int		line_len;
-	int		endian;
-}			t_image;
+	void		*img;
+	char		*addr;
+	int			bpp;
+	int			line_len;
+	int			endian;
+}				t_image;
 
 typedef struct s_draw
 {
-	bool	redraw;
-	int		zoom;
-	int		max_iter;
-	int		x_offset;
-	int		y_offset;
-	int		x_mouse;
-	int		y_mouse;
-}			t_draw;
+	bool		redraw;
+	int			zoom;
+	int			max_iter;
+	int			x_offset;
+	int			y_offset;
+	long double	x_mouse;
+	long double	y_mouse;
+}				t_draw;
 
 typedef struct s_mlx
 {
-	void	*ptr;
-	void	*win;
-	t_image	*img;
-	t_draw	*draw;
-}			t_mlx;
+	void		*ptr;
+	void		*win;
+	t_image		*img;
+	t_draw		*draw;
+}				t_mlx;
 
-void		init(t_mlx *mlx, t_image *image, t_draw *draw);
-int			destroy_and_free(t_mlx *mlx);
-int			key_hook(int key, t_mlx *mlx);
-int			mouse_hook(int key, int x, int y, t_mlx *mlx);
-int			handlebrot(t_mlx *mlx);
-void		load_hooks(t_mlx *mlx);
-void		put_pixel(t_mlx *mlx, t_pixel p);
-void		draw_mandelbrot(t_mlx *mlx);
+void			init(t_mlx *mlx, t_image *image, t_draw *draw);
+int				destroy_and_free(t_mlx *mlx);
+int				key_hook(int key, t_mlx *mlx);
+int				mouse_hook(int key, int x, int y, t_mlx *mlx);
+int				handlebrot(t_mlx *mlx);
+void			load_hooks(t_mlx *mlx);
+void			put_pixel(t_mlx *mlx, t_pixel p);
+void			draw_mandelbrot(t_mlx *mlx);
 
 #endif
